src/server/tasks/dbbench/requirements.txt
mysql-connector-python~=9.3
